+++ changelog.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## [Unreleased]
+### enhancements
+- Remove unnecessary tooltip collision on toolbar items.
+
 ## [1.3.1]
 ### fix
 - disable binding monitors and monitors on apps configurations for now.